+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2.0
+
+* Feature: new `timeoutMs` option to customize the delay
+* Feature: new `link.waitForQueries()` function
+
 # 1.1.0
 
 * Feature: Support fragments on interfaces and unions via a new `fragmentIntrospectionQueryResultData` option.